Enhancement: Refactor Keystone V2 for subservices
- Enhancement: Split out the tenant functionality into its own
  class to take advantage of the subservice functionality.
- BUG: tests are currently broken as the model doesn't seem to be
  quite right. Need to investigate and fix.
